Refactor ProductRepository and ButikEntry for a more generic approach
