+++ src/Contracts/ProductRepository.php
@@ -2,28 +2,8 @@
 
 namespace Jonassiewertsen\Butik\Contracts;
 
-use Illuminate\Support\Collection;
-
 interface ProductRepository
 {
-    public function all(): Collection;
-
-    public function find(string $id): ?self;
-
-    public function findBySlug(string $slug): ?self;
-
-    public function exists(string $slug): bool;
-
-    public function query();
-
-    public function update(array $data);
-
-    public function delete(string $id);
-
-    public function collection(): string;
-
-    public function toArray(): array;
-
     public function stock(): int;
 
     public function tax(): object;
