+++ src/Product/Product.php
@@ -2,81 +2,14 @@
 
 namespace Jonassiewertsen\Butik\Product;
 
-use Illuminate\Support\Collection;
-use Jonassiewertsen\Butik\Exceptions\ButikProductException;
+use Jonassiewertsen\Butik\Contracts\ProductRepository;
 use Jonassiewertsen\Butik\Support\ButikEntry;
-use Statamic\Facades\Entry;
-use Statamic\Facades\Site;
 
-class Product extends ButikEntry
+class Product extends ButikEntry implements ProductRepository
 {
-    public function all(): Collection
-    {
-        return collect($this->query()->all());
-    }
-
-    public function find(string $id): ?self
-    {
-        if (! $this->entry = Entry::find($id)) {
-            return null;
-        }
-
-        $this->defineAttributes();
-
-        return $this;
-    }
-
-    public function findBySlug(string $slug): ?self
-    {
-        $this->entry = Entry::query()
-            ->where('site', Site::current()->handle())
-            ->where('collection', $this->collection())
-            ->where('slug', $slug)
-            ->first();
-
-        if (! $this->entry) {
-            return null;
-        }
-
-        $this->defineAttributes();
-
-        return $this;
-    }
-
-    public function exists(string $slug): bool
-    {
-        return (bool) Entry::find($slug);
-    }
-
-    public function query()
-    {
-        return Entry::whereCollection($this->collection());
-    }
-
-    public function fresh(): self
-    {
-        return $this->find($this->id);
-    }
-
-    public function update(array $data): bool
-    {
-        $data = array_merge($this->data, $data);
-
-        return $this->entry->data($data)->save();
-    }
-
-    public function delete(string $id): bool
-    {
-        if (! $product = $this->find($id)) {
-            throw ButikProductException::cantDeleteNonExistingProduct($id);
-        }
-
-        return $this->find($id)->entry->delete();
-    }
-
     public function collection(): string
     {
-        return 'products'; // TODO: Pull from config
+        return 'products'; // TODO: Get from config
     }
 
     public function stock(): int
@@ -101,6 +34,6 @@ class Product extends ButikEntry
 
     public function toArray(): array
     {
-        //
+        return []; // TODO: Implement
     }
 }
